pyproject.toml
[project]
name = "iliad-account"
version = "0.1.0"
description = ""
authors = [
    {name = "Gianluca",email = "gianluca@gurutech.it"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[project.scripts]
iliad_account = "iliad_account:main"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
exclude = '''
/(
    \.git
  | \.venv
  | \.env
)/
'''
